+++ packages/ecommerce-app-base/src/__mocks__/index.ts
@@ -1,1 +1,3 @@
 export { products, productsList } from './products';
+export { makeSdkMock } from './mockSdk';
+export { mockContentTypes } from './mockContentTypes';